dbt-cloud-orchestration/downstream/pyproject.toml
[project]
name = "downstream"
requires-python = ">=3.10,<3.15"
version = "0.1.0"
dependencies = [
    "dagster>=1.12.8",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
    "pytest>=7.0.0",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/downstream"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/workspace.yaml",
]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "downstream"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E402"]  # Allow imports after load_dotenv()

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["downstream"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
